Libraries/MediaGenerationKitPy/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "media-generation-kit"
version = "0.1.0"
description = "Python client for Draw Things MediaGenerationKit remote and cloud generation."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "UNLICENSED" }
authors = [
  { name = "Draw Things" }
]
dependencies = [
  "fpzip>=1.2",
  "flatbuffers>=24.3.25",
  "grpcio>=1.62",
  "numpy>=1.24",
  "Pillow>=10",
  "protobuf>=4.21",
]

[project.optional-dependencies]
test = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["src/media_generation_kit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
